Added the playlist detail thus an object.

+++ app/src/main/kotlin/taiwan/no1/app/ssfm/models/data/IDataStore.kt
@@ -106,7 +106,7 @@ interface IDataStore {
 
     fun removePlaylist(entity: PlaylistEntity): Observable<Boolean>
 
-    fun getPlaylistItems(playlistId: Int = -1): Observable<List<PlaylistItemEntity>>
+    fun getPlaylistItems(playlistId: Long = -1): Observable<List<PlaylistItemEntity>>
 
     fun addPlaylistItem(entity: PlaylistItemEntity): Observable<Boolean>
