+++ app/src/main/kotlin/taiwan/no1/app/ssfm/models/data/remote/RemoteDataStore.kt
@@ -213,7 +213,7 @@ class RemoteDataStore constructor(private val context: Context) : IDataStore {
 
     override fun removePlaylist(entity: PlaylistEntity): Observable<Boolean> = TODO()
 
-    override fun getPlaylistItems(playlistId: Int): Observable<List<PlaylistItemEntity>> = TODO()
+    override fun getPlaylistItems(playlistId: Long): Observable<List<PlaylistItemEntity>> = TODO()
 
     override fun addPlaylistItem(entity: PlaylistItemEntity): Observable<Boolean> = TODO()
