+++ app/src/main/kotlin/taiwan/no1/app/ssfm/models/entities/PlaylistItemEntity.kt
@@ -12,7 +12,7 @@ import taiwan.no1.app.ssfm.models.entities.lastfm.BaseEntity
  */
 @Table(database = MusicDatabase::class, allFields = true)
 data class PlaylistItemEntity(@PrimaryKey(autoincrement = true) var id: Long = 0,
-                              var playlist_id: Int = -1,  // History list id is 65535.
+                              var playlist_id: Long = -1,  // History list id is 65535.
                               var music_id: Int = -1,
                               var is_offline: Boolean = false,
                               var track_uri: String = "",
